(tui) font_sz unit mixup
THough the font-size (and thus cell-size) setup was changed to use mm
rather than PT size directly, the internal caching structure and the
_setup config was truncating to size_t, switched to float.

+++ src/shmif/arcan_shmif_tui.h
@@ -103,7 +103,8 @@ struct tui_settings {
 	uint8_t alpha;
 	float ppcm;
 	int hint;
-	size_t font_sz;
+
+	float font_sz; /* mm */
 	size_t cell_w, cell_h;
 
 /* either using strings or pre-opened fonts */
